backend/app/requirements.txt
# Core API
fastapi==0.115.9
uvicorn==0.34.2
python-multipart==0.0.20
pydantic==2.11.4
pydantic-settings==2.9.1

# Document Processing
pdf2image==1.17.0
pytesseract==0.3.13
pdfplumber==0.11.6
Pillow==11.2.1
pypdfium2==4.30.1

# Vector DB & Embeddings
faiss-cpu==1.11.0  # Change to faiss-gpu if using NVIDIA GPUs

sentence-transformers==4.1.0

# LLM & AI
groq==0.24.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.42
transformers==4.51.3
torch==2.7.0  # Use torch==2.7.0+cu118 for CUDA 11.8

# Utilities
python-dotenv==1.1.0
requests==2.32.3
numpy==2.2.5
tqdm==4.67.1
orjson==3.10.18

# Missing but required packages from your original list
opentelemetry-api==1.33.0
opentelemetry-instrumentation-fastapi==0.54b0
httpx==0.28.1
backoff==2.2.1
typing-extensions==4.13.2